Wrote system test for PoldiAutoCorrelation
The test performs the analysis on two samples (6903, 6904) and compares the results to the output of the original fortran program.
Since the results are not exactly the same because of numerical differences and some minor bugfixes on the way, a linear fit
between expectation and result is performed and slope/intercept are checked for acceptable values.